backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "homecare-hub-backend"
version = "1.0.0"
description = "HomeCare Hub FastAPI Backend"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pydantic[email]",
    "psycopg[binary]>=3.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[pbkdf2]>=1.7.4",
    "python-multipart>=0.0.9",
    "anthropic>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
